OSPFv3 works correctly with Authentication Trailer
This implementation assumes that optional LLS data block is not used.
AT-bit is set and checked in options field in OSPFv3 Hello and Database
Description packets. OSPFv3 Authentication Trailer is appended to the
OSPFv3 protocol packet. Cryptographic Sequence Number (CSN) is newly
64bit unsigned number. OSPFv3 header checksum is set to 0 prior to
computation of the OSPFv3 Authentication Trailer message digest.