OSPFv3 auth: read opts from HELLO and DBDES packets
  RFC 5613, 2

  The LLS block MAY be attached to OSPF Hello and Database Description
(DD) packets.  The LLS block MUST NOT be attached to any other OSPF
packet types on generation and MUST be ignored on reception.

  RFC 7166, 2.1

  For OSPFv3 Hello and Database Description packets, the AT-bit
indicates that the AT is present.  For other OSPFv3 packet
types, the OSPFv3 AT-bit setting from the OSPFv3 Hello/Database
Description setting is preserved in the OSPFv3 neighbor data
structure.