RPKI: Channels
If a cache is replaced with another cache during reconfiguring then
old channels have to be flushed and freed.

There used to old flushing channels were trying to wake up by another
cache connection and protocol transition START -> UP and all channels are
trying to go to CS_UP. It caused ASSERT(channel MUST NOT be in CS_FLUSH state)

Note that roa channel may be flushed due of expiration of ROAs and in
this case we doesn't have to free the roa channel.

Work in progress with this...